bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Prathameshdhande22 <prathameshdhande534@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
absl-py = "2.1.0"
annotated-types = "0.6.0"
anyio = "4.2.0"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
astunparse = "1.6.3"
async-lru = "2.0.4"
attrs = "23.2.0"
babel = "2.14.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
cachetools = "5.3.2"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.1"
contourpy = "1.2.0"
cycler = "0.12.1"
debugpy = "1.8.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
exceptiongroup = "1.2.0"
executing = "2.0.1"
fastapi = "0.109.2"
fastjsonschema = "2.19.1"
flatbuffers = "23.5.26"
fonttools = "4.49.0"
fqdn = "1.5.1"
gast = "0.5.4"
grpcio = "1.60.1"
h11 = "0.14.0"
h5py = "3.10.0"
httpcore = "1.0.3"
httpx = "0.26.0"
idna = "3.6"
ipykernel = "6.29.2"
ipython = "8.21.0"
isoduration = "20.11.0"
jedi = "0.19.1"
jinja2 = "3.1.3"
joblib = "1.3.2"
json5 = "0.9.14"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
jupyter-events = "0.9.0"
jupyter-lsp = "2.2.2"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
jupyter-server = "2.12.5"
jupyter-server-terminals = "0.5.2"
jupyterlab = "4.1.1"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.25.3"
keras = "2.15.0"
kiwisolver = "1.4.5"
libclang = "16.0.6"
markdown = "3.5.2"
markupsafe = "2.1.5"
matplotlib = "3.8.3"
matplotlib-inline = "0.1.6"
mistune = "3.0.2"
ml-dtypes = "0.2.0"
nbclient = "0.9.0"
nbconvert = "7.16.0"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
nltk = "3.8.1"
notebook-shim = "0.2.4"
numpy = "1.26.4"
oauthlib = "3.2.2"
opt-einsum = "3.3.0"
overrides = "7.7.0"
packaging = "23.2"
pandas = "2.2.0"
pandocfilters = "1.5.1"
parso = "0.8.3"
pillow = "10.2.0"
pip-upgrade = "0.0.6"
platformdirs = "4.2.0"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.43"
protobuf = "4.25.3"
psutil = "5.9.8"
pure-eval = "0.2.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.6.1"
pydantic-core = "2.16.2"
pygments = "2.17.2"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-json-logger = "2.0.7"
pytz = "2024.1"
pywin32 = "306"
pywinpty = "2.0.12"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
referencing = "0.33.0"
regex = "2023.12.25"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.18.0"
rsa = "4.9"
scikit-learn = "1.4.0"
scipy = "1.12.0"
seaborn = "0.13.2"
send2trash = "1.8.2"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
stack-data = "0.6.3"
starlette = "0.36.3"
tensorboard = "2.15.2"
tensorboard-data-server = "0.7.2"
tensorflow = "2.15.0"
tensorflow-estimator = "2.15.0"
tensorflow-intel = "2.15.0"
tensorflow-io-gcs-filesystem = "0.31.0"
termcolor = "2.4.0"
terminado = "0.18.0"
threadpoolctl = "3.3.0"
tinycss2 = "1.2.1"
tomli = "2.0.1"
tornado = "6.4"
tqdm = "4.66.2"
traitlets = "5.14.1"
types-python-dateutil = "2.8.19.20240106"
typing-extensions = "4.9.0"
tzdata = "2024.1"
uri-template = "1.3.0"
urllib3 = "2.2.0"
uvicorn = "0.27.1"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
werkzeug = "3.0.1"
wrapt = "1.14.1"
pyarrow = "^15.0.0"
websockets = "^12.0"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
